Feat: reworked create-password, optimized auth routes to use one authLayout
